Seperate DockerHelper class into it's own file.
